Lock entire venv
