+++ crates/puffin-cli/src/commands/pip_install.rs
@@ -96,6 +96,7 @@ pub(crate) async fn pip_install(
         "Using Python interpreter: {}",
         venv.python_executable().display()
     );
+    let _lock = venv.lock()?;
 
     // Determine the set of installed packages.
     let site_packages =
